删除serial库，解决了Eigen3.4.0无法编译的问题

+++ src/racecar_algorithm/rf2o_laser_odometry/src/CLaserOdometry2D.cpp
@@ -768,8 +768,8 @@ void CLaserOdometry2D::performWarping()
         //Very close pixel
         if (std::abs(std::round(uwarp) - uwarp) < 0.05f)
         {
-          range_warped[image_level](round(uwarp)) += range_w;
-          wacu(std::round(uwarp)) += 1.f;
+          range_warped[image_level]((int)round(uwarp)) += range_w;
+          wacu((int)std::round(uwarp)) += 1.f;
         }
         else
         {
